MARCMP Day 1

+++ Robot-2016/src/Robot2016/Commands/Shooter/SetShooterDeckAngle.h
@@ -19,7 +19,7 @@ public:
 	void Interrupted();
 	virtual void stopAll() override;
 
-	int angle_;
+	float angle_;
 
 private:
 	static const bool BANG_BANG = false;
